add secure fast queue/delete sequence by deleting batches on time and creating an additional deleting state

+++ library/src/main/java/com/novoda/downloadmanager/DownloadBatch.java
@@ -8,6 +8,7 @@ import java.util.List;
 import java.util.Map;
 
 import static com.novoda.downloadmanager.DownloadBatchStatus.Status.DELETED;
+import static com.novoda.downloadmanager.DownloadBatchStatus.Status.DELETING;
 import static com.novoda.downloadmanager.DownloadBatchStatus.Status.DOWNLOADED;
 import static com.novoda.downloadmanager.DownloadBatchStatus.Status.DOWNLOADING;
 import static com.novoda.downloadmanager.DownloadBatchStatus.Status.ERROR;
@@ -51,79 +52,139 @@ class DownloadBatch {
     }
 
     void download() {
-        DownloadBatchStatus.Status status = downloadBatchStatus.status();
+        String rawBatchId = downloadBatchStatus.getDownloadBatchId().rawId();
+        Log.v("start download " + rawBatchId + ", status: " + downloadBatchStatus.status());
 
-        if (status == DELETED) {
-            deleteBatchIfNeeded();
-            notifyCallback(downloadBatchStatus);
+        if (shouldAbortStartingBatch(connectionChecker, callback, downloadBatchStatus, downloadsBatchPersistence)) {
+            Log.v("abort starting download " + rawBatchId);
             return;
         }
 
-        if (status == PAUSED) {
-            notifyCallback(downloadBatchStatus);
-            return;
-        }
-
-        if (connectionNotAllowedForDownload(status)) {
-            processNetworkError();
-            notifyCallback(downloadBatchStatus);
-            return;
-        }
-
-        if (status != DOWNLOADED) {
-            downloadBatchStatus.markAsDownloading(downloadsBatchPersistence);
-            notifyCallback(downloadBatchStatus);
-        }
+        markAsDownloadingIfNeeded(downloadBatchStatus, downloadsBatchPersistence, callback);
 
         if (totalBatchSizeBytes == 0) {
-            totalBatchSizeBytes = getTotalSize(downloadFiles);
+            totalBatchSizeBytes = getTotalSize(downloadFiles, downloadBatchStatus);
         }
 
-        if (totalBatchSizeBytes <= ZERO_BYTES) {
-            deleteBatchIfNeeded();
-            processNetworkError();
-            notifyCallback(downloadBatchStatus);
+        if (shouldAbortAfterGettingTotalBatchSize(downloadBatchStatus, downloadsBatchPersistence, callback, totalBatchSizeBytes)) {
+            Log.v("abort after getting total batch size download " + rawBatchId);
             return;
         }
 
         for (DownloadFile downloadFile : downloadFiles) {
-            if (connectionNotAllowedForDownload(status)) {
-                downloadBatchStatus.markAsWaitingForNetwork(downloadsBatchPersistence);
-                notifyCallback(downloadBatchStatus);
+            if (batchCannotContinue(downloadBatchStatus, connectionChecker, downloadsBatchPersistence, callback)) {
                 break;
             }
             downloadFile.download(fileDownloadCallback);
-            if (batchCannotContinue()) {
-                break;
-            }
         }
 
-        if (networkError()) {
-            processNetworkError();
+        if (networkError(downloadBatchStatus)) {
+            processNetworkError(downloadBatchStatus, callback, downloadsBatchPersistence);
         }
 
-        notifyCallback(downloadBatchStatus);
-        deleteBatchIfNeeded();
+        deleteBatchIfNeeded(downloadBatchStatus, downloadsBatchPersistence, callback);
+        notifyCallback(callback, downloadBatchStatus);
         callbackThrottle.stopUpdates();
+        Log.v("end download " + rawBatchId);
     }
 
-    private void deleteBatchIfNeeded() {
-        if (downloadBatchStatus.status() == DELETED) {
-            downloadsBatchPersistence.deleteAsync(downloadBatchStatus.getDownloadBatchId());
+    private static boolean shouldAbortStartingBatch(ConnectionChecker connectionChecker,
+                                                    DownloadBatchStatusCallback callback,
+                                                    InternalDownloadBatchStatus downloadBatchStatus,
+                                                    DownloadsBatchPersistence downloadsBatchPersistence) {
+        DownloadBatchStatus.Status status = downloadBatchStatus.status();
+
+        if (status == DELETED) {
+            return true;
+        }
+
+        if (status == DELETING) {
+            deleteBatchIfNeeded(downloadBatchStatus, downloadsBatchPersistence, callback);
+            notifyCallback(callback, downloadBatchStatus);
+            return true;
+        }
+
+        if (status == PAUSED) {
+            notifyCallback(callback, downloadBatchStatus);
+            return true;
+        }
+
+        if (connectionNotAllowedForDownload(connectionChecker, status)) {
+            processNetworkError(downloadBatchStatus, callback, downloadsBatchPersistence);
+            notifyCallback(callback, downloadBatchStatus);
+            return true;
+        }
+
+        return false;
+    }
+
+    private static boolean shouldAbortAfterGettingTotalBatchSize(InternalDownloadBatchStatus downloadBatchStatus,
+                                                                 DownloadsBatchPersistence downloadsBatchPersistence,
+                                                                 DownloadBatchStatusCallback callback,
+                                                                 long totalBatchSizeBytes) {
+        if (downloadBatchStatus.status() == DELETING) {
+            deleteBatchIfNeeded(downloadBatchStatus, downloadsBatchPersistence, callback);
+            notifyCallback(callback, downloadBatchStatus);
+            return true;
+        }
+
+        if (totalBatchSizeBytes <= ZERO_BYTES) {
+            processNetworkError(downloadBatchStatus, callback, downloadsBatchPersistence);
+            notifyCallback(callback, downloadBatchStatus);
+            return true;
+        }
+
+        return false;
+    }
+
+    private static void markAsDownloadingIfNeeded(InternalDownloadBatchStatus downloadBatchStatus,
+                                                  DownloadsBatchPersistence downloadsBatchPersistence,
+                                                  DownloadBatchStatusCallback callback) {
+        if (downloadBatchStatus.status() != DOWNLOADED) {
+            downloadBatchStatus.markAsDownloading(downloadsBatchPersistence);
+            notifyCallback(callback, downloadBatchStatus);
         }
     }
 
-    private void processNetworkError() {
-        if (downloadBatchStatus.status() == DELETED) {
+    private static void deleteBatchIfNeeded(InternalDownloadBatchStatus downloadBatchStatus,
+                                            DownloadsBatchPersistence downloadsBatchPersistence,
+                                            DownloadBatchStatusCallback callback) {
+        if (downloadBatchStatus.status() == DELETING) {
+            downloadsBatchPersistence.delete(downloadBatchStatus);
+            Log.v("deleteBatchIfNeeded markAsDeleted: " + downloadBatchStatus.getDownloadBatchId().rawId());
+            downloadBatchStatus.markAsDeleted();
+            notifyCallback(callback, downloadBatchStatus);
+        }
+    }
+
+    private static void processNetworkError(InternalDownloadBatchStatus downloadBatchStatus,
+                                            DownloadBatchStatusCallback callback,
+                                            DownloadsBatchPersistence downloadsBatchPersistence) {
+        if (downloadBatchStatus.status() == DELETING) {
             return;
         }
         downloadBatchStatus.markAsWaitingForNetwork(downloadsBatchPersistence);
-        notifyCallback(downloadBatchStatus);
+        notifyCallback(callback, downloadBatchStatus);
         DownloadsNetworkRecoveryCreator.getInstance().scheduleRecovery();
     }
 
-    private boolean connectionNotAllowedForDownload(DownloadBatchStatus.Status status) {
+    private static boolean connectionNotAllowedForDownload(ConnectionChecker connectionChecker, DownloadBatchStatus.Status status) {
         return !connectionChecker.isAllowedToDownload() && status != DOWNLOADED;
+    }
+
+    private static boolean batchCannotContinue(InternalDownloadBatchStatus downloadBatchStatus,
+                                               ConnectionChecker connectionChecker,
+                                               DownloadsBatchPersistence downloadsBatchPersistence,
+                                               DownloadBatchStatusCallback callback) {
+        DownloadBatchStatus.Status status = downloadBatchStatus.status();
+
+        if (connectionNotAllowedForDownload(connectionChecker, status)) {
+            downloadBatchStatus.markAsWaitingForNetwork(downloadsBatchPersistence);
+            notifyCallback(callback, downloadBatchStatus);
+            return true;
+        } else {
+            return status == ERROR || status == DELETING || status == PAUSED || status == WAITING_FOR_NETWORK;
+        }
     }
 
     private final DownloadFile.Callback fileDownloadCallback = new DownloadFile.Callback() {
@@ -149,7 +210,7 @@ class DownloadBatch {
         }
     };
 
-    private boolean networkError() {
+    private static boolean networkError(InternalDownloadBatchStatus downloadBatchStatus) {
         DownloadBatchStatus.Status status = downloadBatchStatus.status();
         if (status == WAITING_FOR_NETWORK) {
             return true;
@@ -162,12 +223,7 @@ class DownloadBatch {
         return false;
     }
 
-    private boolean batchCannotContinue() {
-        DownloadBatchStatus.Status status = downloadBatchStatus.status();
-        return status == ERROR || status == DELETED || status == PAUSED || status == WAITING_FOR_NETWORK;
-    }
-
-    private long getBytesDownloadedFrom(Map<DownloadFileId, Long> fileBytesDownloadedMap) {
+    private static long getBytesDownloadedFrom(Map<DownloadFileId, Long> fileBytesDownloadedMap) {
         long bytesDownloaded = 0;
         for (Map.Entry<DownloadFileId, Long> entry : fileBytesDownloadedMap.entrySet()) {
             bytesDownloaded += entry.getValue();
@@ -175,19 +231,22 @@ class DownloadBatch {
         return bytesDownloaded;
     }
 
-    private void notifyCallback(InternalDownloadBatchStatus downloadBatchStatus) {
+    private static void notifyCallback(DownloadBatchStatusCallback callback, InternalDownloadBatchStatus downloadBatchStatus) {
         if (callback != null) {
-            callback.onUpdate(downloadBatchStatus);
+            callback.onUpdate(downloadBatchStatus.copy());
         }
     }
 
-    private long getTotalSize(List<DownloadFile> downloadFiles) {
+    private static long getTotalSize(List<DownloadFile> downloadFiles, InternalDownloadBatchStatus downloadBatchStatus) {
         long totalBatchSize = 0;
         for (DownloadFile downloadFile : downloadFiles) {
-            if (downloadBatchStatus.status() == DELETED) {
+            DownloadBatchStatus.Status status = downloadBatchStatus.status();
+            if (status == DELETING || status == DELETED) {
                 return 0;
             }
-            Log.v("DownloadBatch.getTotalSize() " + downloadBatchStatus.getDownloadBatchId().rawId() + ", status: " + downloadBatchStatus.status());
+
+            Log.v("batch id: " + downloadBatchStatus.getDownloadBatchId().rawId() + ", status: " + status + ", file: " + downloadFile.id().rawId());
+
             long totalFileSize = downloadFile.getTotalSize();
             if (totalFileSize == 0) {
                 return 0;
@@ -204,7 +263,7 @@ class DownloadBatch {
             return;
         }
         downloadBatchStatus.markAsPaused(downloadsBatchPersistence);
-        notifyCallback(downloadBatchStatus);
+        notifyCallback(callback, downloadBatchStatus);
 
         for (DownloadFile downloadFile : downloadFiles) {
             downloadFile.pause();
@@ -228,21 +287,33 @@ class DownloadBatch {
             return;
         }
         downloadBatchStatus.markAsQueued(downloadsBatchPersistence);
-        notifyCallback(downloadBatchStatus);
+        notifyCallback(callback, downloadBatchStatus);
         for (DownloadFile downloadFile : downloadFiles) {
             downloadFile.resume();
         }
     }
 
     void delete() {
-        if (downloadBatchStatus.status() == PAUSED) {
-            downloadsBatchPersistence.deleteAsync(downloadBatchStatus.getDownloadBatchId());
+        DownloadBatchStatus.Status status = downloadBatchStatus.status();
+        if (status == DELETING || status == DELETED) {
+            return;
         }
 
-        downloadBatchStatus.markAsDeleted();
-        notifyCallback(downloadBatchStatus);
+        Log.v("delete batch " + downloadBatchStatus.getDownloadBatchId().rawId() + ", mark as deleting from " + downloadBatchStatus.status());
+        downloadBatchStatus.markAsDeleting();
+        notifyCallback(callback, downloadBatchStatus);
+
         for (DownloadFile downloadFile : downloadFiles) {
             downloadFile.delete();
+        }
+
+        if (status == PAUSED || status == DOWNLOADED) {
+            Log.v("delete paused or downloaded batch " + downloadBatchStatus.getDownloadBatchId().rawId());
+            downloadsBatchPersistence.deleteAsync(downloadBatchStatus, downloadBatchId -> {
+                Log.v("delete paused or downloaded mark as deleted: " + downloadBatchId.rawId());
+                downloadBatchStatus.markAsDeleted();
+                notifyCallback(callback, downloadBatchStatus);
+            });
         }
     }
 
@@ -266,6 +337,17 @@ class DownloadBatch {
 
     void persistAsync() {
         downloadsBatchPersistence.persistAsync(
+                downloadBatchStatus.getDownloadBatchTitle(),
+                downloadBatchStatus.getDownloadBatchId(),
+                downloadBatchStatus.status(),
+                downloadFiles,
+                downloadBatchStatus.downloadedDateTimeInMillis(),
+                downloadBatchStatus.notificationSeen()
+        );
+    }
+
+    void persist() {
+        downloadsBatchPersistence.persist(
                 downloadBatchStatus.getDownloadBatchTitle(),
                 downloadBatchStatus.getDownloadBatchId(),
                 downloadBatchStatus.status(),
