+++ library/src/main/java/com/novoda/downloadmanager/RoomDownloadsPersistence.java
@@ -2,8 +2,6 @@ package com.novoda.downloadmanager;
 
 import android.content.Context;
 
-import com.novoda.notils.logger.simple.Log;
-
 import java.util.ArrayList;
 import java.util.List;
 
@@ -37,7 +35,6 @@ final class RoomDownloadsPersistence implements DownloadsPersistence {
 
     @Override
     public void persistBatch(DownloadsBatchPersisted batchPersisted) {
-        Log.v("start persistBatch " + batchPersisted.downloadBatchId().rawId());
         RoomBatch roomBatch = new RoomBatch();
         roomBatch.id = batchPersisted.downloadBatchId().rawId();
         roomBatch.status = batchPersisted.downloadBatchStatus().toRawValue();
@@ -46,7 +43,6 @@ final class RoomDownloadsPersistence implements DownloadsPersistence {
         roomBatch.notificationSeen = batchPersisted.notificationSeen();
 
         database.roomBatchDao().insert(roomBatch);
-        Log.v("end persistBatch " + batchPersisted.downloadBatchId().rawId());
     }
 
     @Override
@@ -70,7 +66,6 @@ final class RoomDownloadsPersistence implements DownloadsPersistence {
 
     @Override
     public void persistFile(DownloadsFilePersisted filePersisted) {
-        Log.v("start persistFile " + filePersisted.downloadBatchId().rawId());
         RoomFile roomFile = new RoomFile();
         roomFile.totalSize = filePersisted.totalFileSize();
         roomFile.batchId = filePersisted.downloadBatchId().rawId();
@@ -81,7 +76,6 @@ final class RoomDownloadsPersistence implements DownloadsPersistence {
         roomFile.persistenceType = filePersisted.filePersistenceType().toRawValue();
 
         database.roomFileDao().insert(roomFile);
-        Log.v("end persistFile " + filePersisted.downloadBatchId().rawId());
     }
 
     @Override
@@ -116,27 +110,24 @@ final class RoomDownloadsPersistence implements DownloadsPersistence {
 
     @Override
     public void delete(DownloadBatchId downloadBatchId) {
-        Log.v("start delete " + downloadBatchId.rawId());
         RoomBatch roomBatch = database.roomBatchDao().load(downloadBatchId.rawId());
+        if (roomBatch == null) {
+            return;
+        }
         database.roomBatchDao().delete(roomBatch);
-        Log.v("end delete " + downloadBatchId.rawId());
     }
 
     @Override
     public void update(DownloadBatchId downloadBatchId, DownloadBatchStatus.Status status) {
-        Log.v("start update " + downloadBatchId.rawId() + ", status: " + status);
         RoomBatch roomBatch = database.roomBatchDao().load(downloadBatchId.rawId());
         roomBatch.status = status.toRawValue();
         database.roomBatchDao().update(roomBatch);
-        Log.v("end update " + downloadBatchId.rawId() + ", status: " + status);
     }
 
     @Override
     public void update(DownloadBatchId downloadBatchId, boolean notificationSeen) {
-        Log.v("start update notification " + downloadBatchId.rawId());
         RoomBatch roomBatch = database.roomBatchDao().load(downloadBatchId.rawId());
         roomBatch.notificationSeen = notificationSeen;
         database.roomBatchDao().update(roomBatch);
-        Log.v("end update  notification" + downloadBatchId.rawId());
     }
 }
